+++ CHANGES.md
@@ -12,6 +12,7 @@ Changes in 8.8
 - Date select increase years options from +5 to +20 in Date.php
 - JS Raise height by 1 submenu item so we stay above browser URL in warehouse.js
 - Add Min. and Max. GPA to Last row in Grades.fnc.php & ReportCards.fnc.php
+- Add Class Rank to Last row in Grades.fnc.php & ReportCards.fnc.php
 
 Changes in 8.7
 --------------